% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condMI.R
\name{condMI}
\alias{condMI}
\title{Calculate conditional mutual information}
\usage{
condMI(idx, statemat, norm = TRUE)
}
\arguments{
\item{idx}{rows of the state matrix for which to compute mutual information}

\item{statemat}{state matrix where each individual is a column and elements are observed states}

\item{norm}{Should the overall result be normalized to the total entropy}
}
\value{
mutual information for each individual in nats
}
\description{
Calculate the conditional mutual information for a subset of the state matrix
}
\examples{
\dontrun{
  test = condMI(idx,statemat)
}
}
\keyword{internal}
