% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_day_break.R
\name{day_break}
\alias{day_break}
\title{Find day break}
\usage{
day_break(posixcttime, dbtime = "07:00:00",
  timezone = extract_time_zone(Sys.time()), prev = F)
}
\arguments{
\item{posixcttime}{time}

\item{dbtime}{String denoting time of daybreak in 24-hour format. Default is '07:00:00' (7AM)}

\item{timezone}{String denoting timezone. Default is system timezone.}

\item{prev}{If TRUE, find previous daybreak. If FALSE, find next daybreak.}
}
\value{
Time of previous or next daybreaks
}
\description{
Find the day break given time in POSIXct format
}
\examples{
# Find next daybreak
next_day_break(Sys.time())
prev_day_break(Sys.time())
}
