% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcaSystemReadHDF5.R
\name{downsample_function}
\alias{downsample_function}
\title{downsample HCA timeseries}
\usage{
downsample_function(subjdat)
}
\arguments{
\item{subjdat}{subject data timeseries}
}
\value{
Return downsampled timeseries
}
\description{
go through timeseries for a subject and downsample to each row per second
}
