% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_start_time.R
\name{find_start_time}
\alias{find_start_time}
\title{Find start time for HCA file}
\usage{
find_start_time(x)
}
\arguments{
\item{x}{input file}
}
\value{
Start time in POSIXct format
}
\description{
Finds the start time for an HCA HDF5 file using the filename string
}
\examples{
infile = '1547228190839_000001_arbC8350154@LT-PINK_0000003600000.hdf5'
find_start_time(infile)
}
\keyword{internal}
