% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_to_hcadf.R
\name{hca_to_dist_df}
\alias{hca_to_dist_df}
\title{Obtain distance and speed}
\usage{
hca_to_dist_df(subjectdatalist, strict.input = F)
}
\arguments{
\item{subjectdatalist}{input HCA object}

\item{strict.input}{Error if input is not of class HCA}
}
\value{
Dataframes with columns for interval distance travelled (mm) and interval speed (mm/s)
}
\description{
Obtain distance and speed from HCA object
}
\examples{
\dontrun{
  hca_to_dist_df(subjectdatalist)
}
}
