% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_to_hcadf.R
\name{hca_to_hcadf}
\alias{hca_to_hcadf}
\title{Convert HCA object to HCA dataframe}
\usage{
hca_to_hcadf(subjectdatalist, dim = 1, strict.input = F)
}
\arguments{
\item{subjectdatalist}{input HCA object}

\item{dim}{number of dimensions to consider. Can only be 1 or 2. If 1, then only mouse position along the length of the cage is returned.}

\item{strict.input}{Error if input is not of class HCA}
}
\value{
Dataframes with columns for time and subject's position
}
\description{
Convert HCA object to HCA dataframe with column for each subject's 1D position
}
\examples{
\dontrun{
  hca_to_hcadf(subjectdatalist)
}
}
