% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_activity.R
\name{HCAdf_to_activity}
\alias{HCAdf_to_activity}
\title{Home cage activity summary}
\usage{
HCAdf_to_activity(HCAdf, aggre.time = 900, norm = TRUE, strict.input = F)
}
\arguments{
\item{HCAdf}{HCAdf object}

\item{aggre.time}{timebins in seconds}

\item{norm}{If TRUE, antenna breaks per hour will be returned. If FALSE, total number of breaks in timebin is returned.}

\item{strict.input}{Error if input is not of class HCAdf}
}
\value{
activity summary
}
\description{
Computes activity (number of transitions) for each timebin of a HCAdf object
}
\examples{
\dontrun{
  test = HCAdf_to_activity(asdfsadf)
}
}
