% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_mi.R
\name{HCAdf_to_MI}
\alias{HCAdf_to_MI}
\title{Mutual Information summary}
\usage{
HCAdf_to_MI(HCAdf, aggre.time = 900, norm = TRUE, strict.input = F)
}
\arguments{
\item{HCAdf}{HCAdf object}

\item{aggre.time}{timebins in seconds}

\item{norm}{flag to compute entropy}

\item{strict.input}{Error if input is not of class HCAdf}
}
\value{
mutual information summary
}
\description{
Computes nutual information for each timebin of a HCAdf object
}
\examples{
\dontrun{
  test = HCAdf_to_entropy(asdfsadf)
}
}
