% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcaReadHDF5.R
\name{hcaReadHDF5}
\alias{hcaReadHDF5}
\title{Read a HCA HDF5 file using rhdf5}
\usage{
hcaReadHDF5(i, subjectIDs = NULL, start.time = NULL, downsample = T)
}
\arguments{
\item{i}{input HDF5 file}

\item{subjectIDs}{List of subjects IDs. If NULL, the subjects are automatically determined}

\item{start.time}{The start time of HCA HDF5 file. If NULL, it is automatically determined.}

\item{downsample}{If TRUE, Aggregate data by second (Default). If FALSE, don't aggregate.}
}
\value{
List of dataframes, one for each subjects.
}
\description{
Read an HCA HDF5 file and return a list of dataframes, one for each subjects
}
\examples{
\dontrun{
  hcaReadHDF5('1547743375205_000095_arbC8123198@LT-GREEN_0000085500200.hdf5')
}
}
