% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcaReadHDF5series.R
\name{hcaReadHDF5series}
\alias{hcaReadHDF5series}
\title{Read many HCA HDF5 files}
\usage{
hcaReadHDF5series(hdf5files, subjectIDs = NULL, parallel = NULL)
}
\arguments{
\item{hdf5files}{input HDF5 files}

\item{subjectIDs}{List of subjects IDs. If NULL, the subjects are automatically determined.}

\item{parallel}{number of local cores to use. If NULL, process is run in serial.}
}
\value{
List of dataframes, one for each subjects.
}
\description{
Read a series of HCA HDF5 files and return a list of dataframes, one for each subjects
}
\examples{
\dontrun{
  hdf5files = c(
                 '1547743375205_000093_arbC8123198@LT-GREEN_0000083700200.hdf5'
                 '1547743375205_000094_arbC8123198@LT-GREEN_0000084600200.hdf5',
                 '1547743375205_000095_arbC8123198@LT-GREEN_0000085500200.hdf5',
               )
  hcaGetHDF5series(hdf5files)
}
}
