% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcaSystemReadHDF5.R
\name{hcaSystemReadHDF5}
\alias{hcaSystemReadHDF5}
\title{Read an HCA HDF5 file and return a list of dataframes, one for each subjects}
\usage{
hcaSystemReadHDF5(
  i,
  subjectIDs = NULL,
  start.time = NULL,
  downsample = TRUE,
  h5ls_path = NULL,
  h5dump_path = NULL
)
}
\arguments{
\item{i}{input HDF5 file}

\item{subjectIDs}{List of subjects IDs. If NULL, the subjects are automatically determined}

\item{start.time}{The start time of HCA HDF5 file. If NULL, it is automatically determined.}

\item{downsample}{If TRUE, Aggregate data by second (Default). If FALSE, don't aggregate.}

\item{h5ls_path}{Path to h5ls utility. If NULL (Default), path is assumed to be in current environment,}

\item{h5dump_path}{Path to h5dump utility. If NULL (Default), path is assumed to be in current environment.}
}
\value{
List of dataframes, one for each subjects.
}
\description{
Read an HCA HDF5 file and return a list of dataframes, one for each subjects
}
\examples{
\dontrun{
  hcaReadHDF5('1547743375205_000095_arbC8123198@LT-GREEN_0000085500200.hdf5')
}
}
