% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcaSystemReadHDF5.R
\name{hcaSystemReadHDF5_defense_func}
\alias{hcaSystemReadHDF5_defense_func}
\title{defence for hcaSystemReadHDF5 inputs}
\usage{
hcaSystemReadHDF5_defense_func(i, h5ls_path, h5dump_path)
}
\arguments{
\item{i}{input HDF5 file}

\item{h5ls_path}{Path to h5ls utility.}

\item{h5dump_path}{Path to h5dump utility.}
}
\description{
chech arguments for hcaSystemReadHDF5 function and stop if problematic
}
