% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcaSystemReadHDF5.R
\name{hcaSystemReadHDF5_subj_defense_func}
\alias{hcaSystemReadHDF5_subj_defense_func}
\title{defence for hcaSystemReadHDF5 subjects}
\usage{
hcaSystemReadHDF5_subj_defense_func(i, h5ls_path)
}
\arguments{
\item{i}{input HDF5 file}

\item{h5ls_path}{Path to h5ls utility.}
}
\description{
chech input file for hcaSystemReadHDF5 function and stop if no subjects
}
