% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeCageEntropy.R
\name{homeCageEntropy}
\alias{homeCageEntropy}
\title{Calculate home cage entropy}
\usage{
homeCageEntropy(idx, statemat, norm = TRUE)
}
\arguments{
\item{idx}{rows of the state matrix for which to compute entropy}

\item{statemat}{state matrix where each individual is a column and elements are observed states}

\item{norm}{Should the overall result be normalized to the total number of states (6)}
}
\value{
entropy for each individual in nats
}
\description{
Calculate the home cage entropy for a subset of the state matrix
}
\examples{
\dontrun{
  test = homeCageEntropy(idx,statemat)
}
}
\keyword{internal}
