% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcamode.R
\name{Mode}
\alias{Mode}
\title{Find mode}
\usage{
Mode(x, breaktie = T)
}
\arguments{
\item{x}{input vector}

\item{breaktie}{if TRUE (default), then tie is broken with the first value. If FALSE, all tied values are returned.}
}
\value{
most frequent value in x
}
\description{
Determine the most frequent value
}
\examples{
x = c(1,1,1,2,2,2,3,4)
Mode(x) 
Mode(x,breaktie=F) 
}
\keyword{internal}
