% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcaReadHDF5.R
\name{no_message_h5ls}
\alias{no_message_h5ls}
\title{suppress console messages from rhdf5::h5ls}
\usage{
no_message_h5ls(...)
}
\arguments{
\item{...}{arguments to be passed to h5ls}
}
\value{
Return from h5ls
}
\description{
Read an HCA HDF5 file and return a list of dataframes, one for each subjects
}
