% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hca.R
\name{plot.HCA}
\alias{plot.HCA}
\title{Plot HCA}
\usage{
\method{plot}{HCA}(subjectdatalist, idxs = NULL)
}
\arguments{
\item{subjectdatalist}{HCA object}

\item{idxs}{Which index is plotted. If NULL, all plots are made.}
}
\value{
Plot of subjects horizontal position
}
\description{
Plot method for HCA object
}
