% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hcaent.R
\name{plot.HCAentropy}
\alias{plot.HCAentropy}
\title{Plot HCA entropy}
\usage{
\method{plot}{HCAentropy}(hcaentdf, idxs = NULL)
}
\arguments{
\item{hcaentdf}{HCAentropy object}

\item{idxs}{Which subject is plotted. If NULL, all plots are made.}
}
\value{
Plot of subjects' entropy timeseries
}
\description{
Plot method for HCAentropy object
}
