% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hcami.R
\name{plot.HCAmi}
\alias{plot.HCAmi}
\title{Plot HCA MI}
\usage{
\method{plot}{HCAmi}(hcamidf, idxs = NULL)
}
\arguments{
\item{hcamidf}{HCAmi object}

\item{idxs}{Which subject is plotted. If NULL, all plots are made.}
}
\value{
Plot of subjects' mutual information timeseries
}
\description{
Plot method for HCAmi object
}
