% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsubjfunc.R
\name{readsubjfunc}
\alias{readsubjfunc}
\title{Subjects in HDF5 files}
\usage{
readsubjfunc(i)
}
\arguments{
\item{i}{HDF5 filename}
}
\value{
List of subjects
}
\description{
List subjects in HCA HDF5 files
}
\examples{
\dontrun{
  readsubjfunc('1547743375205_000095_arbC8123198@LT-GREEN_0000085500200.hdf5')
}
}
\keyword{internal}
