% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transEnt.R
\name{TransEnt}
\alias{TransEnt}
\title{Calculate transfer entropy}
\usage{
TransEnt(idx, statemat, norm = TRUE)
}
\arguments{
\item{idx}{rows of the state matrix for which to compute mutual information}

\item{statemat}{state matrix where each individual is a column and elements are observed states}

\item{norm}{Should the overall result be normalized to the total entropy}
}
\value{
transfer entropy for each pairwise-individuals in nats
}
\description{
Calculate the transfer entropy for a subset of the state matrix
}
\examples{
\dontrun{
  test = TransEnt(idx,statemat)
}
}
\keyword{internal}
