% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvec_to_dayvec.R
\name{tvec_to_dayvec}
\alias{tvec_to_dayvec}
\title{Annotate time with day or night}
\usage{
tvec_to_dayvec(
  posix_time,
  daystart = times("07:00:01"),
  nightstart = times("19:00:00")
)
}
\arguments{
\item{posix_time}{vector of times given in POSIXct or POSIXt format}

\item{daystart}{time when the light phase starts}

\item{nightstart}{time when the dark phase starts}
}
\value{
vector of day or night
}
\description{
Converts a vector of POSIXct or POSIXt into day or night
}
\examples{
\dontrun{
  test = tvec_to_dayvec(asdfsadf)
}
}
\keyword{internal}
