% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_TE.R
\name{HCAdf_to_TE}
\alias{HCAdf_to_TE}
\title{Transfer entropy summary}
\usage{
HCAdf_to_TE(HCAdf, lagtime = 900, aggre.time = 900, norm = TRUE,
  int.pairs = F, strict.input = F)
}
\arguments{
\item{HCAdf}{HCAdf object}

\item{aggre.time}{timebins in seconds}

\item{norm}{flag to compute entropy}

\item{strict.input}{Error if input is not of class HCAdf}
}
\value{
mutual information summary
}
\description{
Computes transfer entropy for each timebin of a HCAdf object
}
\examples{
\dontrun{
  test = HCAdf_to_entropy(asdfsadf)
}
}
