% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hcami.R, R/plot_hcate.R
\name{plot.HCAmi}
\alias{plot.HCAmi}
\alias{plot.HCAmi}
\title{Plot HCA MI}
\usage{
\method{plot}{HCAmi}(hcamidf, idxs = NULL)

\method{plot}{HCAte}(hcamidf, idxs = NULL)
}
\arguments{
\item{hcamidf}{HCAmi object}

\item{idxs}{Which subject is plotted. If NULL, all plots are made.}

\item{hcatedf}{HCAte object}

\item{idxs}{Which subject (or pairs) is plotted. If NULL, all plots are made.}
}
\value{
Plot of subjects' mutual information timeseries

Plot of subjects' (or pairs) transfer entropy timeseries
}
\description{
Plot method for HCAmi object

Plot method for HCAte object
}
